+++ sentry/src/main/java/io/sentry/Hub.java
@@ -30,6 +30,7 @@ public final class Hub implements IHub {
   private final @NotNull TracesSampler tracesSampler;
   private final @NotNull Map<Throwable, Pair<WeakReference<ISpan>, String>> throwableToSpan =
       Collections.synchronizedMap(new WeakHashMap<>());
+  private final @NotNull TransactionPerformanceCollector transactionPerformanceCollector;
 
   public Hub(final @NotNull SentryOptions options) {
     this(options, createRootStackItem(options));
@@ -44,6 +45,7 @@ public final class Hub implements IHub {
     this.tracesSampler = new TracesSampler(options);
     this.stack = stack;
     this.lastEventId = SentryId.EMPTY_ID;
+    this.transactionPerformanceCollector = new TransactionPerformanceCollector(options);
 
     // Integrations will use this Hub instance once registered.
     // Make sure Hub ready to be used then.
@@ -730,7 +732,8 @@ public final class Hub implements IHub {
               waitForChildren,
               idleTimeout,
               trimEnd,
-              transactionFinishedCallback);
+              transactionFinishedCallback,
+              transactionPerformanceCollector);
 
       // The listener is called only if the transaction exists, as the transaction is needed to
       // stop it